chore: clean up code, refactor and use User Settings for Legend component
