+++ app/javascript/react/types/userStates.ts
@@ -3,4 +3,5 @@ export enum UserSettings {
   ModalView = "MODAL_VIEW",
   CalendarView = "CALENDAR_VIEW",
   SessionListView = "SESSION_LIST_VIEW",
+  MapLegendView = "MAP_LEGEND_VIEW",
 }
